Replaced old synth with TinySoundFont

+++ Source/AlphaTab/Audio/Synth/IAlphaSynth.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using AlphaTab.Audio.Synth.Midi;
-using AlphaTab.Audio.Synth.Synthesis;
 
 namespace AlphaTab.Audio.Synth
 {
@@ -34,7 +33,7 @@ namespace AlphaTab.Audio.Synth
         }
 
         /// <summary>
-        /// Gets or sets the loging level. 
+        /// Gets or sets the loging level.
         /// </summary>
         AlphaTab.Util.LogLevel LogLevel
         {
@@ -89,7 +88,7 @@ namespace AlphaTab.Audio.Synth
 
         /// <summary>
         /// Gets or sets the range of the song that should be played. Set this to null
-        /// to play the whole song. 
+        /// to play the whole song.
         /// </summary>
         PlaybackRange PlaybackRange
         {
@@ -98,7 +97,7 @@ namespace AlphaTab.Audio.Synth
         }
 
         /// <summary>
-        /// Gets or sets whether the playback should automatically restart after it finished. 
+        /// Gets or sets whether the playback should automatically restart after it finished.
         /// </summary>
         bool IsLooping
         {
@@ -113,7 +112,7 @@ namespace AlphaTab.Audio.Synth
 
         /// <summary>
         /// Starts the playback if possible
-        /// </summary>  
+        /// </summary>
         void Play();
 
         /// <summary>
@@ -138,7 +137,7 @@ namespace AlphaTab.Audio.Synth
         void LoadSoundFont(byte[] data);
 
         /// <summary>
-        /// Loads the given midi file structure. 
+        /// Loads the given midi file structure.
         /// </summary>
         /// <param name="midi"></param>
         void LoadMidiFile(MidiFile midi);
@@ -167,8 +166,8 @@ namespace AlphaTab.Audio.Synth
         /// Gets or sets the current and initial volume of the given channel.
         /// </summary>
         /// <param name="channel">The channel number.</param>
-        /// <param name="volume">The volume of of the channel (0.0-3.0)</param>
-        void SetChannelVolume(int channel, double volume);
+        /// <param name="volume">The volume of of the channel (0.0-1.0)</param>
+        void SetChannelVolume(int channel, float volume);
 
         /// <summary>
         /// Gets or sets the current and initial program of the given channel.
@@ -178,47 +177,47 @@ namespace AlphaTab.Audio.Synth
         void SetChannelProgram(int channel, byte program);
 
         /// <summary>
-        /// This event is fired when the player is ready to be interacted with. 
+        /// This event is fired when the player is ready to be interacted with.
         /// </summary>
         event Action Ready;
 
         /// <summary>
-        /// This event is fired when all required data for playback is loaded and ready. 
+        /// This event is fired when all required data for playback is loaded and ready.
         /// </summary>
         event Action ReadyForPlayback;
 
         /// <summary>
-        /// This event is fired when the playback of the whole song finished. 
+        /// This event is fired when the playback of the whole song finished.
         /// </summary>
         event Action<PlaybackFinishedEventArgs> Finished;
 
         /// <summary>
-        /// This event is fired when the SoundFont needed for playback was loaded. 
+        /// This event is fired when the SoundFont needed for playback was loaded.
         /// </summary>
         event Action SoundFontLoaded;
 
         /// <summary>
-        /// This event is fired when the loading of the SoundFont failed. 
+        /// This event is fired when the loading of the SoundFont failed.
         /// </summary>
         event Action<Exception> SoundFontLoadFailed;
 
         /// <summary>
-        /// This event is fired when the Midi file needed for playback was loaded. 
+        /// This event is fired when the Midi file needed for playback was loaded.
         /// </summary>
         event Action MidiLoaded;
 
         /// <summary>
-        /// This event is fired when the loading of the Midi file failed. 
+        /// This event is fired when the loading of the Midi file failed.
         /// </summary>
         event Action<Exception> MidiLoadFailed;
 
         /// <summary>
-        /// This event is fired when the playback state changed. 
+        /// This event is fired when the playback state changed.
         /// </summary>
         event Action<PlayerStateChangedEventArgs> StateChanged;
 
         /// <summary>
-        /// This event is fired when the current playback position of the song changed. 
+        /// This event is fired when the current playback position of the song changed.
         /// </summary>
         event Action<PositionChangedEventArgs> PositionChanged;
     }
